voucher: se termina alta y modificacion de items

+++ lib/model/DTOs/voucher_item_dto.dart
@@ -7,7 +7,7 @@ class VoucherItemDTO extends Deserializable<VoucherItemDTO> {
   String? medicineName;
   double? currentStock;
   String? presentation;
-  int? quantity;
+  double? quantity;
   double? unitPrice;
   //final bool? isFirst;
